Move roughness_rms to statistics.py
